amiquip

Enum Error

pub enum Error {

    UrlParseError {
        source: ParseError,
    },
    TlsFeatureNotEnabled,
    InsecureUrl {
        url: Url,
    },
    UnexpectedSocketClose,
    IoErrorReadingSocket {
        source: Error,
    },
    IoErrorWritingSocket {
        source: Error,
    },
    MalformedFrame,
    UrlNoSocketAddrs {
        url: Url,
    },
    ResolveUrlToSocketAddr {
        url: Url,
        source: Error,
    },
    FailedToConnect {
        url: Url,
        source: Error,
    },
    SpecifyUrlPort {
        url: Url,
    },
    InvalidUrlScheme {
        url: Url,
    },
    UrlMissingDomain {
        url: Url,
    },
    ExtraUrlPathSegments {
        url: Url,
    },
    UrlParseHeartbeat {
        url: Url,
        source: ParseIntError,
    },
    UrlParseChannelMax {
        url: Url,
        source: ParseIntError,
    },
    UrlParseConnectionTimeout {
        url: Url,
        source: ParseIntError,
    },
    UrlInvalidAuthMechanism {
        url: Url,
        mechanism: String,
    },
    UrlUnsupportedParameter {
        url: Url,
        parameter: String,
    },
    CreatePollHandle {
        source: Error,
    },
    RegisterWithPollHandle {
        source: Error,
    },
    DeregisterWithPollHandle {
        source: Error,
    },
    FailedToPoll {
        source: Error,
    },
    TlsHandshake {
        source: Error,
    },
    CreateTlsConnector {
        source: Error,
    },
    UnsupportedAuthMechanism {
        available: String,
        requested: String,
    },
    UnsupportedLocale {
        available: String,
        requested: String,
    },
    FrameMaxTooSmall {
        min: u32,
        requested: u32,
    },
    ConnectionTimeout,
    SaslSecureNotSupported,
    InvalidCredentials,
    MissedServerHeartbeats,
    ServerClosedConnection {
        code: u16,
        message: String,
    },
    ClientClosedConnection,
    ServerClosedChannel {
        channel_id: u16,
        code: u16,
        message: String,
    },
    ClientClosedChannel,
    EventLoopClientDropped,
    EventLoopDropped,
    FrameUnexpected,
    ForkFailed {
        source: Error,
    },
    ExhaustedChannelIds,
    UnavailableChannelId {
        channel_id: u16,
    },
    ClientException,
    ReceivedFrameWithBogusChannelId {
        channel_id: u16,
    },
    IoThreadPanic,
    DuplicateConsumerTag {
        channel_id: u16,
        consumer_tag: String,
    },
    UnknownConsumerTag {
        channel_id: u16,
        consumer_tag: String,
    },
    // some variants omitted

}

Specific error cases returned by amiquip.

Variants

UrlParseError

URL parsing failed.

Fields

source: ParseError

TlsFeatureNotEnabled

A TLS connection was requested (e.g., via URL), but the amiquip was built without TLS support.

InsecureUrl

An insecure URL was supplied to Connection::open, which only allows amqps://... URLs.

Fields

url: Url

UnexpectedSocketClose

The underlying socket was closed.

IoErrorReadingSocket

An I/O error occurred while reading from the socket.

Fields

source: Error

IoErrorWritingSocket

An I/O error occurred while writing from the socket.

Fields

source: Error

MalformedFrame

We received data that could not be parsed as an AMQP frame.

UrlNoSocketAddrs

Failed to resolve a URL into an IP address (or addresses).

Fields

url: Url

ResolveUrlToSocketAddr

Error resolving a URL into an IP address (or addresses).

Fields

url: Url

source: Error

FailedToConnect

Failed to open TCP connection.

Fields

url: Url

source: Error

SpecifyUrlPort

Failed to set the port on a URL.

Fields

url: Url

InvalidUrlScheme

Invalid scheme for URL.

Fields

url: Url

UrlMissingDomain

URL is missing domain.

Fields

url: Url

ExtraUrlPathSegments

URL contains extra path segments.

Fields

url: Url

UrlParseHeartbeat

Could not parse heartbeat parameter of URL.

Fields

url: Url

source: ParseIntError

UrlParseChannelMax

Could not parse channel_max parameter of URL.

Fields

url: Url

source: ParseIntError

UrlParseConnectionTimeout

Could not parse connection_timeout parameter of URL.

Fields

url: Url

source: ParseIntError

UrlInvalidAuthMechanism

Invalid auth mechanism requested in URL.

Fields

url: Url

mechanism: String

UrlUnsupportedParameter

Unsupported URL parameter.

Fields

url: Url

parameter: String

CreatePollHandle

Could not create mio Poll handle.

Fields

source: Error

RegisterWithPollHandle

Could not register descriptor with Poll handle.

Fields

source: Error

DeregisterWithPollHandle

Could not register descriptor with Poll handle.

Fields

source: Error

FailedToPoll

Failed to poll mio Poll handle.

Fields

source: Error

TlsHandshake

The TLS handshake failed.

Fields

source: Error

CreateTlsConnector

Error from underlying TLS implementation.

Fields

source: Error

UnsupportedAuthMechanism

The server does not support the requested auth mechanism.

Fields

available: String

requested: String

UnsupportedLocale

The server does not support the requested locale.

Fields

available: String

requested: String

FrameMaxTooSmall

The requested frame size is smaller than the minimum required by AMQP.

Fields

min: u32

requested: u32

ConnectionTimeout

Timeout occurred while performing the initial TCP connection.

SaslSecureNotSupported

The server requested a Secure/Secure-Ok exchange, which are currently unsupported.

InvalidCredentials

The supplied authentication credentials were not accepted by the server.

MissedServerHeartbeats

The server missed too many successive heartbeats.

ServerClosedConnection

The server closed the connection with the given reply code and text.

Fields

code: u16

message: String

ClientClosedConnection

The client closed the connection.

ServerClosedChannel

The server closed the given channel with the given reply code and text.

Fields

channel_id: u16

code: u16

message: String

ClientClosedChannel

The client closed the channel.

EventLoopClientDropped

The I/O loop attempted to send a message to a caller that did not exist. This indicates either a bug in amiquip or a connection that is in a bad state and in the process of tearing down.

EventLoopDropped

The I/O loop has dropped the sending side of a channel, typically because it has exited due to another error.

FrameUnexpected

We received a valid AMQP frame but not one we expected; e.g., receiving an incorrect response to an AMQP method call.

ForkFailed

Forking the I/O thread failed.

Fields

source: Error

ExhaustedChannelIds

No more channels can be opened because there are already channel_max channels open.

UnavailableChannelId

An explicit channel ID was requested, but that channel is unavailable for use (e.g., because there is another open channel with the same ID).

Fields

channel_id: u16

ClientException

The client sent an AMQP exception to the server and closed the connection.

ReceivedFrameWithBogusChannelId

The server sent frames for a channel ID we don’t know about.

Fields

channel_id: u16

IoThreadPanic

The I/O thread panicked.

DuplicateConsumerTag

The server sent us a consumer tag that is equal to another consumer tag we already have on the same channel.

Fields

channel_id: u16

consumer_tag: String

UnknownConsumerTag

The server sent us a Delivery for a channel we don’t know about.

Fields

channel_id: u16

consumer_tag: String