amiquip

Struct Exchange

pub struct Exchange<'a> { /* private fields */ }

Handle for a declared AMQP exchange.

Implementations

impl Exchange<'_>

pub fn direct(channel: &Channel) -> Exchange<'_>

Construct a handle for the direct exchange on the given channel. This is an entirely local operation; the default exchange (named "") is guaranteed to exist and does not need to be declared.

pub fn name(&self) -> &str

Name of this exchange.

pub fn publish(&self, publish: Publish<'_>) -> Result<()>

Publish a message to this exchange.

pub fn bind_to_source<S: Into<String>>( &self, source: &Exchange<'_>, routing_key: S, arguments: FieldTable, ) -> Result<()>

Synchronously bind this exchange (as destination) to the source exchange with the given routing key and arguments. Exchange-to-exchange binding is a RabbitMQ extension; you can examine the connection’s server properties to see if the current connection supports this feature.

pub fn bind_to_source_nowait<S: Into<String>>( &self, source: &Exchange<'_>, routing_key: S, arguments: FieldTable, ) -> Result<()>

Asynchronously bind this exchange (as destination) to the source exchange with the given routing key and arguments. Exchange-to-exchange binding is a RabbitMQ extension; you can examine the connection’s server properties to see if the current connection supports this feature.

pub fn bind_to_destination<S: Into<String>>( &self, destination: &Exchange<'_>, routing_key: S, arguments: FieldTable, ) -> Result<()>

Synchronously bind this exchange (as source) to the destination exchange with the given routing key and arguments. Exchange-to-exchange binding is a RabbitMQ extension; you can examine the connection’s server properties to see if the current connection supports this feature.

pub fn bind_to_destination_nowait<S: Into<String>>( &self, destination: &Exchange<'_>, routing_key: S, arguments: FieldTable, ) -> Result<()>

Asynchronously bind this exchange (as source) to the destination exchange with the given routing key and arguments. Exchange-to-exchange binding is a RabbitMQ extension; you can examine the connection’s server properties to see if the current connection supports this feature.

pub fn unbind_from_source<S: Into<String>>( &self, source: &Exchange<'_>, routing_key: S, arguments: FieldTable, ) -> Result<()>

Synchronously unbind this exchange (as destination) from the source exchange with the given routing key and arguments. Exchange-to-exchange binding is a RabbitMQ extension; you can examine the connection’s server properties to see if the current connection supports this feature.

pub fn unbind_from_source_nowait<S: Into<String>>( &self, source: &Exchange<'_>, routing_key: S, arguments: FieldTable, ) -> Result<()>

Asynchronously unbind this exchange (as destination) from the source exchange with the given routing key and arguments. Exchange-to-exchange binding is a RabbitMQ extension; you can examine the connection’s server properties to see if the current connection supports this feature.

pub fn unbind_from_destination<S: Into<String>>( &self, destination: &Exchange<'_>, routing_key: S, arguments: FieldTable, ) -> Result<()>

Synchronously unbind this exchange (as source) from the destination exchange with the given routing key and arguments. Exchange-to-exchange binding is a RabbitMQ extension; you can examine the connection’s server properties to see if the current connection supports this feature.

pub fn unbind_from_destination_nowait<S: Into<String>>( &self, destination: &Exchange<'_>, routing_key: S, arguments: FieldTable, ) -> Result<()>

Asynchronously unbind this exchange (as source) from the destination exchange with the given routing key and arguments. Exchange-to-exchange binding is a RabbitMQ extension; you can examine the connection’s server properties to see if the current connection supports this feature.

pub fn delete(self, if_unused: bool) -> Result<()>

Synchronously delete this exchange. If if_unused is true, the exchange will only be deleted if it has no queue bindings; if if_unused is true and the exchange still has queue bindings, the server will close this channel.

pub fn delete_nowait(self, if_unused: bool) -> Result<()>

Asynchronously delete this exchange. If if_unused is true, the exchange will only be deleted if it has no queue bindings; if if_unused is true and the exchange still has queue bindings, the server will close this channel.